fix: undo renaming locker service
