+++ src/background/services/lock/__tests__/lockerService.test.ts
