+++ src/background/services/lock/index.ts
