+++ src/background/services/zkIdentity/index.ts
@@ -3,7 +3,7 @@ import { browser } from "webextension-polyfill-ts";
 
 import BrowserUtils from "@src/background/controllers/browserUtils";
 import HistoryService from "@src/background/services/history";
-import LockerService from "@src/background/services/locker";
+import LockerService from "@src/background/services/lock";
 import NotificationService from "@src/background/services/notification";
 import SimpleStorage from "@src/background/services/storage";
 import { ZkIdentitySemaphore } from "@src/background/services/zkIdentity/protocols/ZkIdentitySemaphore";
